add local development scripts
