+++ package.json
@@ -3,14 +3,19 @@
   "version": "1.0.0",
   "license": "MIT",
   "scripts": {
-    "build:webhooks:original": "ts-node ./scripts/build-webhooks.ts",
-    "build:webhooks": "ts-node ./scripts/build-webhooks-oas3.1.ts"
+    "serve": "http-server --port=8080 --cors=*",
+    "build:webhooks": "ts-node ./scripts/build-webhooks.ts",
+    "dev:webhooks": "ts-node ./scripts/build-webhooks.ts local",
+    "start:webhooks": "yarn serve & yarn watch:webhooks",
+    "watch:webhooks": "nodemon --watch ./models/webhooks --ext yaml --exec yarn dev:webhooks"
   },
   "devDependencies": {
     "@types/js-yaml": "^4.0.0",
     "@types/node": "^14.14.37",
     "chalk": "^4.1.0",
+    "http-server": "^0.12.3",
     "js-yaml": "^4.0.0",
+    "nodemon": "^2.0.7",
     "openapi-merge": "^1.0.29",
     "shelljs": "^0.8.4",
     "ts-node": "^9.1.1",
